requirement.txt
annotated-types==0.7.0
anyio==4.5.2
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
dnspython==2.6.1
elastic-transport==8.13.1
elasticsearch==6.3.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastexcel==0.11.5
idna==3.7
importlib_metadata==8.2.0
inflect==7.3.1
logger==1.4     
more-itertools==10.3.0
mysql-connector-python==9.0.0
numpy==1.24.4
openpyxl==3.1.5
polars==1.3.0
pyarrow==17.0.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.44.0
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.6
XlsxWriter==3.2.0
zipp==3.19.2
uvicorn==0.30.6
toml==0.10.2
opencv-python
Pillow
torch>=1.7
torchvision
tqdm
pandas==2.2.3
pillow==11.1.0
boto3==1.36.22
pika==1.3.2
mysql-connector-python==9.0.0
pymongo==4.10.1
python-multipart==0.0.20
libmagic==1.0
python-magic==0.4.27
Jinja2==3.1.6
sentence-transformers==3.3.1
scikit-learn==1.5.2
huggingface_hub==0.27.1
nltk==3.9.1
streamlit==1.43.0
pymilvus==2.5.11
